refactor: Split version check into package
