+++ internal/deploy/version_check.go
@@ -5,55 +5,27 @@ package deploy
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
-	"fmt"
-	"io"
 	"log/slog"
-	"net/http"
-	"net/url"
-	"os"
 	"runtime"
 	"strings"
 	"time"
 
-	"github.com/blang/semver/v4"
 	"github.com/exasol/exasol-personal/internal/config"
+	"github.com/exasol/exasol-personal/internal/version_check"
 )
 
 const (
-	VersionCheckURLEnvVar = "EXASOL_VERSION_CHECK_URL"
-	VersionCheckCategory  = "Exasol Personal"
-
+	VersionCheckCategory    = "Exasol Personal"
 	VersionCheckLockTimeout = 250 * time.Millisecond
 )
-
-// GetVersionCheckURL resolves the version-check endpoint URL.
-// The endpoint can be overridden for tests and controlled environments.
-func GetVersionCheckURL() string {
-	versionCheckURL := os.Getenv(VersionCheckURLEnvVar)
-	if versionCheckURL == "" {
-		versionCheckURL = DefaultVersionCheckURL
-	}
-
-	return versionCheckURL
-}
-
-// VersionCheckDetails contains platform-specific information for version checking.
-type VersionCheckDetails struct {
-	OperatingSystem string
-	Architecture    string
-	Category        string
-	ClusterIdentity string
-	URL             string
-}
 
 // GetVersionCheckDetails returns platform-specific version check details.
 // The version check URL can be overridden using the EXASOL_VERSION_CHECK_URL environment variable.
 // If deploymentDir is provided and contains a state file, uses its persisted ClusterIdentity.
 // If no valid state file exists (e.g. when called outside a deployment directory),
 // ClusterIdentity is left empty.
-func GetVersionCheckDetails(deployment config.DeploymentDir) *VersionCheckDetails {
+func GetVersionCheckDetails(deployment config.DeploymentDir) *version_check.VersionCheckDetails {
 	operatingSystem := runtime.GOOS
 	arch := runtime.GOARCH
 
@@ -88,124 +60,13 @@ func GetVersionCheckDetails(deployment config.DeploymentDir) *VersionCheckDetail
 		}
 	}
 
-	return &VersionCheckDetails{
+	return &version_check.VersionCheckDetails{
 		OperatingSystem: operatingSystem,
 		Architecture:    arch,
 		Category:        VersionCheckCategory,
 		ClusterIdentity: clusterIdentity,
-		URL:             GetVersionCheckURL(),
+		URL:             version_check.GetVersionCheckURL(),
 	}
-}
-
-// LatestVersionInfo contains information about the latest version.
-type LatestVersionInfo struct {
-	Version         string `json:"version"`
-	Filename        string `json:"filename"`
-	URL             string `json:"url"`
-	Size            int64  `json:"size"`
-	SHA256          string `json:"sha256"`
-	OperatingSystem string `json:"operatingSystem"`
-	Architecture    string `json:"architecture"`
-}
-
-// VersionCheckResponse represents the response from the version check API.
-type VersionCheckResponse struct {
-	LatestVersion LatestVersionInfo `json:"latestVersion"`
-}
-
-// SilentVersionCheckResult reports whether a version check was performed and its outcome.
-type SilentVersionCheckResult struct {
-	Checked         bool
-	UpdateAvailable bool
-	LatestVersion   string
-}
-
-// IsVersionUpdateAvailable reports whether latestVersion is newer than currentVersion.
-func IsVersionUpdateAvailable(currentVersion, latestVersion string) (bool, error) {
-	current, err := semver.Parse(currentVersion)
-	if err != nil {
-		return false, fmt.Errorf(
-			"failed to parse current launcher version %q: %w",
-			currentVersion,
-			err,
-		)
-	}
-	latest, err := semver.Parse(latestVersion)
-	if err != nil {
-		return false, fmt.Errorf(
-			"failed to parse latest launcher version %q: %w",
-			latestVersion,
-			err,
-		)
-	}
-
-	return latest.GT(current), nil
-}
-
-func parseVersionCheckResponse(body io.Reader) (*VersionCheckResponse, error) {
-	result := &VersionCheckResponse{}
-
-	if err := json.NewDecoder(body).Decode(result); err != nil {
-		return nil, fmt.Errorf("failed to parse version check response: %w", err)
-	}
-
-	return result, nil
-}
-
-const versionCheckTimeout = 3 // seconds
-
-// CheckLatestVersion checks for the latest version from the API.
-func CheckLatestVersion(
-	ctx context.Context,
-	details *VersionCheckDetails,
-	currentVersion string,
-) (*VersionCheckResponse, error) {
-	params := url.Values{}
-	params.Add("category", details.Category)
-	params.Add("operatingSystem", details.OperatingSystem)
-	params.Add("architecture", details.Architecture)
-	params.Add("version", currentVersion)
-	if strings.TrimSpace(details.ClusterIdentity) != "" {
-		params.Add("identity", details.ClusterIdentity)
-	}
-
-	requestURL := fmt.Sprintf("%s?%s", details.URL, params.Encode())
-
-	slog.Debug("making version check request", "url", requestURL)
-
-	// Create HTTP client with timeout
-	client := &http.Client{
-		Timeout: versionCheckTimeout * time.Second,
-	}
-
-	// Make the request using the provided context
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, requestURL, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %w", err)
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("failed to make version check request: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		body, _ := io.ReadAll(resp.Body)
-		return nil, fmt.Errorf(
-			"version check request failed with status %d: %s",
-			resp.StatusCode,
-			string(body),
-		)
-	}
-
-	// Parse the response
-	result, err := parseVersionCheckResponse(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	return result, nil
 }
 
 // FetchLatestVersion retrieves version information for the current platform.
@@ -213,9 +74,9 @@ func FetchLatestVersion(
 	ctx context.Context,
 	currentVersion string,
 	deployment config.DeploymentDir,
-) (*VersionCheckResponse, error) {
+) (*version_check.VersionCheckResponse, error) {
 	details := GetVersionCheckDetails(deployment)
-	return CheckLatestVersion(ctx, details, currentVersion)
+	return version_check.CheckLatestVersion(ctx, details, currentVersion)
 }
 
 func MustDoVersionCheck(exasolState *config.ExasolPersonalState) bool {
@@ -252,7 +113,7 @@ func CheckLatestVersionUpdate(
 		return false, "", err
 	}
 	latest := response.LatestVersion.Version
-	available, err := IsVersionUpdateAvailable(currentVersion, latest)
+	available, err := version_check.IsVersionUpdateAvailable(currentVersion, latest)
 	if err != nil {
 		return false, latest, err
 	}
@@ -266,10 +127,10 @@ func PerformSilentVersionCheck(
 	ctx context.Context,
 	deployment config.DeploymentDir,
 	currentVersion string,
-) (SilentVersionCheckResult, error) {
+) (version_check.SilentVersionCheckResult, error) {
 	slog.Debug("begin version update check")
 
-	result := SilentVersionCheckResult{}
+	result := version_check.SilentVersionCheckResult{}
 
 	lockCtx, cancel := context.WithTimeout(ctx, VersionCheckLockTimeout)
 	defer cancel()
